+++ tests/drum/run-drum-tests-in-container.sh
@@ -58,7 +58,6 @@ cd $GIT_ROOT || exit 1
 DONE_PREP_TIME=$(date +%s)
 
 # Running mlops monitoring tests without the datarobot-mlops installed
-
 pytest tests/drum/test_mlops_monitoring.py::TestMLOpsMonitoring::test_drum_monitoring_no_mlops_installed
 TEST_RESULT_NO_MLOPS=$?
 
@@ -66,9 +65,8 @@ pip install \
     --extra-index-url https://artifactory.int.datarobot.com/artifactory/api/pypi/python-all/simple \
     datarobot-mlops
 
-pytest tests/drum/test_units.py \
-       tests/drum/test_custom_model.py \
-       tests/drum/test_mlops_monitoring.py::TestMLOpsMonitoring::test_drum_monitoring_with_mlops_installed\
+pytest tests/drum/test_* \
+       -k "not test_drum_monitoring_no_mlops_installed" \
        --junit-xml="$GIT_ROOT/results_integration.xml"
 
 TEST_RESULT=$?
